% Auto-generated: do not edit by hand
\name{dashCreditCardInput}

\alias{dashCreditCardInput}

\title{DashCreditCardInput component}

\description{

}

\usage{
dashCreditCardInput(children=NULL, id=NULL, acceptedCards=NULL, callback=NULL,
cardCVCInputProps=NULL, cardCVCInputRenderer=NULL,
cardExpiryInputProps=NULL, cardExpiryInputRenderer=NULL,
cardImageClassName=NULL, cardImageStyle=NULL,
cardNumber=NULL, cardNumberInputProps=NULL,
cardNumberInputRenderer=NULL, className=NULL,
containerClassName=NULL, containerStyle=NULL,
customTextLabels=NULL, cvc=NULL, dangerTextClassName=NULL,
dangerTextStyle=NULL, expiry=NULL, fieldClassName=NULL,
fieldStyle=NULL, inputClassName=NULL, inputComponent=NULL,
inputStyle=NULL, invalidClassName=NULL, invalidStyle=NULL,
issuer=NULL, onError=NULL, placeholders=NULL, preview=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{acceptedCards}{Unnamed list. }

\item{callback}{}

\item{cardCVCInputProps}{Named list. }

\item{cardCVCInputRenderer}{}

\item{cardExpiryInputProps}{Named list. }

\item{cardExpiryInputRenderer}{}

\item{cardImageClassName}{Character. }

\item{cardImageStyle}{Named list. }

\item{cardNumber}{Character. }

\item{cardNumberInputProps}{Named list. }

\item{cardNumberInputRenderer}{}

\item{className}{Character. }

\item{containerClassName}{Character. }

\item{containerStyle}{Named list. }

\item{customTextLabels}{Named list. }

\item{cvc}{Character. }

\item{dangerTextClassName}{Character. }

\item{dangerTextStyle}{Named list. }

\item{expiry}{Character. }

\item{fieldClassName}{Character. }

\item{fieldStyle}{Named list. }

\item{inputClassName}{Character. }

\item{inputComponent}{Character. }

\item{inputStyle}{Named list. }

\item{invalidClassName}{Character. }

\item{invalidStyle}{Named list. }

\item{issuer}{Character. }

\item{onError}{}

\item{placeholders}{Named list. }

\item{preview}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

