% Auto-generated: do not edit by hand
\name{dashCreditCards}

\alias{dashCreditCards}

\title{DashCreditCards component}

\description{

}

\usage{
dashCreditCards(id=NULL, acceptedCards=NULL, cvc=NULL, expiry=NULL,
focused=NULL, issuer=NULL, locale=NULL, name=NULL,
number=NULL, placeholders=NULL, preview=NULL)
}

\arguments{
\item{id}{Character. }

\item{acceptedCards}{Unnamed list. }

\item{cvc}{Character | numeric. }

\item{expiry}{Character | numeric. }

\item{focused}{Character. }

\item{issuer}{Character. }

\item{locale}{Named list. }

\item{name}{Character | numeric. }

\item{number}{Character | numeric. }

\item{placeholders}{Named list. }

\item{preview}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

