% Auto-generated: do not edit by hand
\name{dashParallaxTilt}

\alias{dashParallaxTilt}

\title{DashParallaxTilt component}

\description{

}

\usage{
dashParallaxTilt(children=NULL, id=NULL, className=NULL,
flipHorizontally=NULL, flipVertically=NULL,
glareBorderRadius=NULL, glareColor=NULL, glareEnable=NULL,
glareMaxOpacity=NULL, glarePosition=NULL, glareReverse=NULL,
gyroscope=NULL, onEnter=NULL, onLeave=NULL, onMove=NULL,
perspective=NULL, reset=NULL, scale=NULL, style=NULL,
tiltAngleXInitial=NULL, tiltAngleXManual=NULL,
tiltAngleYInitial=NULL, tiltAngleYManual=NULL,
tiltAxis=NULL, tiltEnable=NULL, tiltMaxAngleX=NULL,
tiltMaxAngleY=NULL, tiltReverse=NULL, trackOnWindow=NULL,
transitionEasing=NULL, transitionSpeed=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{flipHorizontally}{Logical. }

\item{flipVertically}{Logical. }

\item{glareBorderRadius}{Character. }

\item{glareColor}{Character. }

\item{glareEnable}{Logical. }

\item{glareMaxOpacity}{Numeric. }

\item{glarePosition}{A value equal to: 'top', 'right', 'bottom', 'left', 'all'. }

\item{glareReverse}{Logical. }

\item{gyroscope}{Logical. }

\item{onEnter}{}

\item{onLeave}{}

\item{onMove}{}

\item{perspective}{Numeric. }

\item{reset}{Logical. }

\item{scale}{Numeric. }

\item{style}{Named list. }

\item{tiltAngleXInitial}{Numeric. }

\item{tiltAngleXManual}{Numeric. }

\item{tiltAngleYInitial}{Numeric. }

\item{tiltAngleYManual}{Numeric. }

\item{tiltAxis}{A value equal to: 'x', 'y'. }

\item{tiltEnable}{Logical. }

\item{tiltMaxAngleX}{Numeric. }

\item{tiltMaxAngleY}{Numeric. }

\item{tiltReverse}{Logical. }

\item{trackOnWindow}{Logical. }

\item{transitionEasing}{Character. }

\item{transitionSpeed}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

